PyTorch_Alexnet_Exploration/requirements.txt
streamlit<=1.32.2
torch>=1.7.0
torchvision>=0.8.1
Pillow==10.2.0
numpy==1.26.4